+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 0.2
 ---
 
-  * Add `[@key]` and `[@name]` attributes.
+  * Add `[@key]`, `[@name]` and `[@default]` attributes.
   * Add support for `Yojson.Safe.json` values.
 
 0.1